Made a large amount of visual fixes and improvements, and other fixes
UI was pretty broken still after iced upgrade so I had to fix a bunch of
formatting, and while I was at it I made a bunch of improvements such as
preventing the images from being cropped (in most cases), hiding the
prev/next buttons when at the first/last element of the community
showcase, and adding a tooltip to the community showcase.
Also updated some links in the consts file.
One last issue remains, which is the performance of the changelog.