Added mechanism to remove old and unneeded cached images
Firstly, a cache version field was added to the Airshipper application
struct to wipe the whole cache when the version is bumped. This prevents
the need to resize the loaded cached images again before returning them
in case the old, full size versions were present instead from an older
version of Airshipper.

Secondly, a new check was added when an rss feed is updated, so any
images in the cache not present in the current set of wanted images are
deleted. This prevents cached images from accumulating over time, as new
posts get added; only the newest 15 are needed.