on the 0.15 release CI behaved the following: check - build - build2 - publish
where all jobs are green. but publish only contained the gitlab_release.
It was ran on the Merge branch release-0.16 into master.
which ran a 2nd time with "latest" and "tag" comment.

The first one with this was prob the merge train which ran on branch master

now we use the thing that already works on gitlab_release on the flatpak and crates.io one